backing out recent changes to optLong, getLong. See #868
